+++ packages/shared/src/modules/users/libs/types/authenticated-user-patch-request-dto.type.ts
@@ -1,6 +1,7 @@
 type AuthenticatedUserPatchRequestDto = {
-	firstName: string;
-	lastName: string;
+	firstName?: string;
+	isVisibleProfile?: boolean;
+	lastName?: string;
 };
 
 export { type AuthenticatedUserPatchRequestDto };